# backend/empresa/target/classes/api.yaml
openapi: "3.0.3"
info:
  title: Biblioteca API
  description: API para gerenciar organizações
  version: 1.0.0
...
paths:
...
  /endereco:
    post:
      summary: Cria um novo endereço
      security:
        - bearerAuth: []
      tags:
        - Endereço
      operationId: createendereco
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/enderecoModel'
      responses:
        '200':
          description: Endereço criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/enderecoModel'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Endereço não criado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Erro interno."
    get:
      summary: Busca Enderecos
      security:
        - bearerAuth: []
      tags:
        - Perfil
      operationId: findendereco
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/enderecoModel'
  /endereco/{id}:
    put:
      summary: Atualiza um endereço existente
      operationId: updateendereco
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/enderecoModel'
      responses:
        '200':
          description: Endereço atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/enderecoModel'
        '404':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Endereço não encontrado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Erro interno."
    get:
      summary: Detalha um endereço
      security:
        - bearerAuth: []
      tags:
        - Endereço
      operationId: detailendereco
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/enderecoModel'
    delete:
      summary: Deleta um endereço existente
      operationId: deleteendereco
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Endereço excluído com sucesso
        '404':
          description: Endereço não encontrado
        '500':
          description: Erro interno do servidor
...
  /estado:
    get:
      summary: Busca Estados
      security:
        - bearerAuth: []
      tags:
        - Estado
      operationId: findEstado
      parameters:
        - name: nome
          in: query
          required: false
          schema:
            type: string
        - name: pais_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EstadoModel'
  /estado/{id}:
    get:
      summary: Detalha um Estado
      security:
        - bearerAuth: []
      tags:
        - Estado
      operationId: detailEstado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoModel'
...
  /geolocalizacao:
    post:
      summary: Obter coordenadas de um endereço
      security:
        - bearerAuth: []
      tags:
        - Geolocalização
      operationId: getCoordenadas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/enderecoModel'
      responses:
        '200':
          description: Coordenadas obtidas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/enderecoModel'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Coordenadas não obtidas."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Erro interno."
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # Especifica o formato do token, pode ser omitido
  schemas:
...
    enderecoModel:
      type: object
      properties:
        id:
          type: integer
        cidade_id:
          $ref: '#/components/schemas/CidadeModel'
        cep:
          type: string
        logradouro:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        latitude:
          type: number
        longitude:
            type: number
        usuario_id:
          type: integer
...
    CidadeModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        estado_id:
          $ref: '#/components/schemas/EstadoModel'
    EstadoModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        pais_id:
          $ref: '#/components/schemas/PaisModel'
    PaisModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
...