# backend/empresa/target/generated-sources/openapi/src/main/java/com/dasad/empresa/api/GeolocalizacaoApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dasad.empresa.api;

import com.dasad.empresa.model.Createendereco500Response;
import com.dasad.empresa.model.EnderecoModel;
import com.dasad.empresa.model.GetCoordenadas400Response;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-03-15T21:23:23.313933800Z[Europe/Lisbon]", comments = "Generator version: 7.9.0")
@Validated
@Tag(name = "Geolocalização", description = "Operações relacionadas a geolocalização")
public interface GeolocalizacaoApi {

    /**
     * POST /geolocalizacao : Obter coordenadas de um endereço
     *
     * @param enderecoModel  (required)
     * @return Coordenadas obtidas com sucesso (status code 200)
     *         or Requisição inválida (status code 400)
     *         or Erro interno (status code 500)
     */
    @Operation(
        operationId = "getCoordenadas",
        summary = "Obter coordenadas de um endereço",
        tags = { "Geolocalização" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Coordenadas obtidas com sucesso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = EnderecoModel.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requisição inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetCoordenadas400Response.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro interno", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Createendereco500Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/geolocalizacao",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<EnderecoModel> getCoordenadas(
        @Parameter(name = "EnderecoModel", description = "", required = true) @Valid @RequestBody EnderecoModel enderecoModel
    );

}